requirements.txt
email
dnspython
